refactor order helper
