+++ decidim-proposals/app/helpers/decidim/proposals/proposal_order_helper.rb
@@ -1,0 +1,15 @@
+module Decidim
+  module Proposals
+    # Simple helpers to handle proposals ordering
+    module ProposalOrderHelper
+      # Returns the options the user will see to order proposals
+      def order_fields
+        @order_fields ||= begin
+          order_fields = [:random, :recent]
+          order_fields << :most_voted if current_settings.votes_enabled?
+          order_fields
+        end
+      end
+    end
+  end
+end
